fix minified output. remove cjs output
