+++ .eslintrc.json
@@ -1,0 +1,29 @@
+{
+  "root": true,
+
+  "env": {
+    "browser": true
+  },
+
+  "parser": "@typescript-eslint/parser",
+
+  "extends": [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:@typescript-eslint/recommended-requiring-type-checking",
+    "prettier"
+  ],
+
+  "plugins": ["@typescript-eslint"],
+
+  "globals": {
+    "__statics": true,
+    "process": true
+  },
+
+  // add your custom rules here
+  "rules": {
+    "@typescript-eslint/default-param-last": "off",
+    "@typescript-eslint/strict-boolean-expressions": 0
+  }
+}
