+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tauri-plugin-sql-api",  
+  "name": "tauri-plugin-sql-api",
   "version": "0.1.0",
   "description": "API wrapper for the Tauri Plugin for SQL databases",
   "license": "Apache-2.0 OR MIT",
@@ -18,21 +18,10 @@
   "files": [
     "webview-dist"
   ],
-  "tsup": {
-    "entry": [
-      "./webview-dist/index.ts"
-    ],
-    "sourcemap": true,
-    "clean": true,
-    "dts": true
-  },
   "scripts": {
-    "build": "run-p lint build:esm",
-    "build:esm": "tsup webview-src/index.ts --format='esm,cjs' --dts --out-dir webview-dist ",
-    "lint": "run-s lint:*",
-    "lint:webview": "eslint --ext ts \"./webview-src/**/*.ts\"",
-    "lint:rust": "cargo clippy --features sqlite",
-    "lint-fix": "eslint --fix --ext ts \"./src/**/*.ts\"",
+    "build": "rollup -c ./webview-src/rollup.config.js",
+    "lint": "eslint webview-src --ext ts",
+    "lint-fix": "eslint webview-src --ext ts --fix",
     "prepublishOnly": "yarn build",
     "pretest": "yarn build"
   },
@@ -41,20 +30,16 @@
     "tslib": "2.3.1"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "5.10.0",
-    "@typescript-eslint/parser": "5.10.0",
-    "eslint": "8.7.0",
+    "@rollup/plugin-node-resolve": "13.1.3",
+    "@rollup/plugin-typescript": "8.3.0",
+    "@typescript-eslint/eslint-plugin": "5.10.1",
+    "@typescript-eslint/parser": "5.10.1",
+    "eslint": "8.8.0",
     "eslint-config-prettier": "8.3.0",
-    "eslint-config-standard-with-typescript": "21.0.1",
-    "eslint-plugin-import": "2.25.4",
-    "eslint-plugin-lodash-template": "0.19.0",
-    "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "4.0.0",
-    "eslint-plugin-promise": "6.0.0",
-    "eslint-plugin-security": "1.4.0",
-    "npm-run-all": "4.1.5",
     "prettier": "2.5.1",
-    "tsup": "5.11.11",
-    "typescript": "4.5.4"
+    "rollup": "2.66.1",
+    "rollup-plugin-terser": "7.0.2",
+    "typescript": "4.5.5"
   }
 }
