+++ webview-src/index.ts
@@ -32,11 +32,11 @@ export default class Database {
    * A static initializer which connects to the underlying database
    * and returns a `Database` instance once a connecion to the database
    * is established.
-   * 
+   *
    * # Sqlite
-   * 
+   *
    * The path is relative to `tauri::api::path::BaseDirectory::App` and must start with `sqlite:`.
-   * 
+   *
    * ```ts
    * const db = await Database.load("sqlite:test.db");
    * ```
@@ -53,9 +53,9 @@ export default class Database {
    * A static initializer which synchronously returns an instance of
    * the Database class while deferring the actual database connection
    * until the first invokation or selection on the database.
-   * 
+   *
    * # Sqlite
-   * 
+   *
    * The path is relative to `tauri::api::path::BaseDirectory::App` and must start with `sqlite:`.
    *
    * ```ts
@@ -78,7 +78,7 @@ export default class Database {
    * );
    * ```
    */
-  async execute(query: string, bindValues?: any[]): Promise<QueryResult> {
+  async execute(query: string, bindValues?: unknown[]): Promise<QueryResult> {
     return await invoke<[number, number]>('plugin:sql|execute', {
       db: this.path,
       query,
@@ -97,7 +97,7 @@ export default class Database {
    * );
    * ```
    */
-  async select<T>(query: string, bindValues?: any[]): Promise<T> {
+  async select<T>(query: string, bindValues?: unknown[]): Promise<T> {
     return await invoke('plugin:sql|select', {
       db: this.path,
       query,
