+++ webview-src/tsconfig.json
@@ -12,10 +12,7 @@
     "forceConsistentCasingInFileNames": true,
     "skipLibCheck": true,
     "declaration": true,
-
-    "outDir": "./webview-dist",
-    "declarationDir": "./webview-dist"
+    "declarationDir": "../webview-dist"
   },
-  "include": ["webview-src"],
   "exclude": ["node_modules"]
 }
